refactor: extract shared client logic into _BaseClient
Move all non-I/O logic (URL building, request preparation, response
parsing, presigned URL/POST generation) into a shared _BaseClient class.
Both Client and AsyncClient now inherit from it, reducing duplication
from ~1400 lines across two files to ~400 in the base + ~160 each.
