+++ src/signurlarity/aio/client.py
@@ -1,20 +1,16 @@
+"""Async S3 client for presigned URL generation and basic S3 operations."""
+
 from __future__ import annotations
 
-from typing import Any, Mapping, Optional
-from urllib.parse import urlparse
+from typing import Any
 
 import httpx
 
-from ..exceptions import (
-    BucketAlreadyExistsError,
-    BucketAlreadyOwnedByYouError,
-    NoSuchBucketError,
-    PresignError,
-)
-from ..presigner import S3Presigner
+from .._base import _BaseClient
+from ..exceptions import PresignError
 
 
-class AsyncClient:
+class AsyncClient(_BaseClient):
     """Async S3 client for generating presigned URLs and performing S3 operations.
 
     This is a lightweight, boto3-compatible async client that focuses on presigned URL
@@ -99,107 +95,20 @@ class AsyncClient:
         endpoint_url: str,
         aws_access_key_id: str,
         aws_secret_access_key: str,
-        httpx_max_connections: Optional[int] = None,
+        httpx_max_connections: int | None = None,
     ):
-        self.endpoint_url = endpoint_url
-        self.aws_access_key_id = aws_access_key_id
-        self.aws_secret_access_key = aws_secret_access_key
-
-        # Extract region from endpoint URL
-        self.region = self._extract_region(endpoint_url)
-
-        # Initialize the presigner
-        self._presigner = S3Presigner(
-            access_key=aws_access_key_id,
-            secret_key=aws_secret_access_key,
-            region=self.region,
+        super().__init__(
             endpoint_url=endpoint_url,
+            aws_access_key_id=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+            httpx_max_connections=httpx_max_connections,
         )
-
-        # Initialize HTTP client for connection pooling
-        limits = httpx._config.DEFAULT_LIMITS
-        if httpx_max_connections:
-            limits.max_connections = httpx_max_connections
-        self._http_client = httpx.AsyncClient(limits=limits)
-
-    def _extract_region(self, endpoint_url: str) -> str:
-        """Extract AWS region from endpoint URL.
-
-        Args:
-            endpoint_url: S3 endpoint URL
-
-        Returns:
-            AWS region string (defaults to 'us-east-1')
-
-        """
-        parsed = urlparse(endpoint_url)
-        hostname = parsed.hostname or parsed.netloc
-
-        # Parse patterns like:
-        # - s3.amazonaws.com -> us-east-1
-        # - s3.us-west-2.amazonaws.com -> us-west-2
-        # - bucket.s3.eu-central-1.amazonaws.com -> eu-central-1
-        parts = hostname.split(".")
-        if "s3" in parts:
-            idx = parts.index("s3")
-            if idx + 1 < len(parts) and parts[idx + 1] != "amazonaws":
-                return parts[idx + 1]
-
-        return "us-east-1"
-
-    def _build_request_url(
-        self, bucket: str, key: str | None = None
-    ) -> tuple[str, str, dict[str, str]]:
-        """Build request URL and headers for S3 operations.
-
-        Args:
-            bucket: S3 bucket name
-            key: Optional object key
-
-        Returns:
-            Tuple of (base_url, path, headers)
-
-        """
-        scheme = "https"
-        if self.endpoint_url:
-            parsed = urlparse(self.endpoint_url)
-            scheme = parsed.scheme or "https"
-            # For custom endpoints, use path-style URLs
-            actual_host = parsed.netloc
-            if key:
-                path = f"/{bucket}/{key}"
-            else:
-                path = f"/{bucket}"
-            base_url = f"{scheme}://{actual_host}"
-            headers = {"host": actual_host}
-        else:
-            # For AWS endpoints, use virtual-hosted style
-            actual_host = self._presigner._get_host(bucket)
-            path = f"/{key}" if key else "/"
-            base_url = f"{scheme}://{actual_host}"
-            headers = {"host": actual_host}
-
-        return base_url, path, headers
-
-    def _build_query_string(self, query_params: dict[str, str]) -> str:
-        """Build query string from parameters.
-
-        Args:
-            query_params: Dictionary of query parameters
-
-        Returns:
-            Query string (without leading '?')
-
-        """
-        return "&".join(f"{k}={v}" for k, v in query_params.items())
+        self._http_client = httpx.AsyncClient(limits=self._limits)
 
     async def _execute_request(
         self, method: str, url: str, headers: dict[str, str], body: bytes = b""
     ) -> httpx.Response:
-        """Execute an HTTP request using connection pooling.
-
-        Uses the instance's httpx.AsyncClient for connection pooling, reusing
-        the same HTTP connection across multiple requests for better performance.
+        """Execute an HTTP request using async connection pooling.
 
         Args:
             method: HTTP method (GET, PUT, HEAD, DELETE, etc.)
@@ -262,7 +171,6 @@ class AsyncClient:
             https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3/client/generate_presigned_post.html
 
         Example:
-            >>> # Simple POST with ACL
             >>> post_data = await client.generate_presigned_post(
             ...     Bucket="mybucket",
             ...     Key="myfile.txt",
@@ -270,56 +178,20 @@ class AsyncClient:
             ...     Conditions=[["content-length-range", 0, 1048576]],
             ...     ExpiresIn=3600,
             ... )
-            >>> # Use post_data['url'] and post_data['fields'] for upload
-
-            >>> # POST with metadata and content restrictions
-            >>> post_data = await client.generate_presigned_post(
-            ...     Bucket="uploads",
-            ...     Key="photos/vacation.jpg",
-            ...     Fields={
-            ...         "Content-Type": "image/jpeg",
-            ...         "x-amz-meta-photographer": "john-doe",
-            ...         "acl": "private",
-            ...     },
-            ...     Conditions=[
-            ...         ["content-length-range", 1024, 5242880],  # 1KB to 5MB
-            ...         ["eq", "$Content-Type", "image/jpeg"],
-            ...     ],
-            ...     ExpiresIn=900,  # 15 minutes
-            ... )
-            >>> # Upload file asynchronously from Python
-            >>> import httpx
-            >>> async with httpx.AsyncClient() as http_client:
-            ...     with open("vacation.jpg", "rb") as f:
-            ...         files = {"file": f.read()}
-            ...         response = await http_client.post(
-            ...             post_data["url"],
-            ...             data=post_data["fields"],
-            ...             files={"file": files["file"]},
-            ...         )
-            >>> print(response.status_code)  # 204 on success
 
         """
-        if not Bucket:
-            raise PresignError("Missing required parameter 'Bucket'")
-        if not Key:
-            raise PresignError("Missing required parameter 'Key'")
-
-        try:
-            return self._presigner.generate_presigned_post(
-                bucket=Bucket,
-                key=Key,
-                fields=Fields,
-                conditions=Conditions,
-                expires=ExpiresIn,
-            )
-        except ValueError as e:
-            raise PresignError(str(e)) from e
+        return self._generate_presigned_post(
+            Bucket=Bucket,
+            Key=Key,
+            Fields=Fields,
+            Conditions=Conditions,
+            ExpiresIn=ExpiresIn,
+        )
 
     async def generate_presigned_url(
         self,
         ClientMethod: str,
-        Params: Mapping[str, Any],
+        Params: Any,
         ExpiresIn: int = 3600,
         HttpMethod: str = "",
     ) -> str:
@@ -343,89 +215,19 @@ class AsyncClient:
             https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3/client/generate_presigned_url.html
 
         Example:
-            >>> # Generate URL for downloading a file (GET)
             >>> url = await client.generate_presigned_url(
             ...     "get_object",
             ...     Params={"Bucket": "mybucket", "Key": "documents/report.pdf"},
-            ...     ExpiresIn=3600,  # Valid for 1 hour
-            ... )
-            >>> # Share the URL or use it with any async HTTP client
-            >>> import httpx
-            >>> async with httpx.AsyncClient() as http_client:
-            ...     response = await http_client.get(url)
-            ...     with open("downloaded_report.pdf", "wb") as f:
-            ...         f.write(response.content)
-
-            >>> # Generate URL for uploading a file (PUT)
-            >>> upload_url = await client.generate_presigned_url(
-            ...     "put_object",
-            ...     Params={"Bucket": "mybucket", "Key": "uploads/newfile.txt"},
-            ...     ExpiresIn=900,  # Valid for 15 minutes
-            ... )
-            >>> # Upload file asynchronously using the URL
-            >>> import httpx
-            >>> async with httpx.AsyncClient() as http_client:
-            ...     with open("local_file.txt", "rb") as f:
-            ...         response = await http_client.put(upload_url, content=f.read())
-            >>> print(response.status_code)  # 200 on success
-
-            >>> # Generate multiple URLs concurrently
-            >>> import asyncio
-            >>> urls = await asyncio.gather(
-            ...     client.generate_presigned_url(
-            ...         "get_object", {"Bucket": "b1", "Key": "k1"}
-            ...     ),
-            ...     client.generate_presigned_url(
-            ...         "get_object", {"Bucket": "b2", "Key": "k2"}
-            ...     ),
-            ...     client.generate_presigned_url(
-            ...         "get_object", {"Bucket": "b3", "Key": "k3"}
-            ...     ),
+            ...     ExpiresIn=3600,
             ... )
 
         """
-        if Params is None:
-            Params = {}
-
-        # Extract bucket and key from params
-        bucket = Params.get("Bucket")
-        key = Params.get("Key")
-
-        if not bucket:
-            raise PresignError("Missing required parameter 'Bucket' in Params")
-        if not key:
-            raise PresignError("Missing required parameter 'Key' in Params")
-
-        # Determine HTTP method from ClientMethod if not explicitly provided
-        if not HttpMethod:
-            HttpMethod = self._client_method_to_http_method(ClientMethod)
-
-        try:
-            return self._presigner.generate_presigned_url(
-                bucket=bucket, key=key, method=HttpMethod, expires=ExpiresIn
-            )
-        except ValueError as e:
-            raise PresignError(str(e)) from e
-
-    def _client_method_to_http_method(self, client_method: str) -> str:
-        """Map boto3 client method name to HTTP method.
-
-        Args:
-            client_method: S3 client method name (e.g., 'get_object')
-
-        Returns:
-            HTTP method string (GET, PUT, DELETE, etc.)
-
-        """
-        method_map = {
-            "get_object": "GET",
-            "put_object": "PUT",
-            "delete_object": "DELETE",
-            "head_object": "HEAD",
-            "list_objects": "GET",
-            "list_objects_v2": "GET",
-        }
-        return method_map.get(client_method.lower(), "GET")
+        return self._generate_presigned_url(
+            ClientMethod=ClientMethod,
+            Params=Params,
+            ExpiresIn=ExpiresIn,
+            HttpMethod=HttpMethod,
+        )
 
     async def __aenter__(self):
         """Async context manager entry point."""
@@ -440,7 +242,7 @@ class AsyncClient:
         """Ensure HTTP client is closed."""
         await self._http_client.aclose()
 
-    async def head_bucket(self, Bucket: str, **kwargs):
+    async def head_bucket(self, Bucket: str, **kwargs) -> dict[str, Any]:
         """Check if a bucket exists and is accessible.
 
         Performs a HEAD request to the bucket to verify existence and access permissions.
@@ -466,60 +268,11 @@ class AsyncClient:
             'us-west-2'
 
         """
-        if not Bucket:
-            raise PresignError("Missing required parameter 'Bucket'")
-
-        # Build the request
-        base_url, path, headers = self._build_request_url(Bucket)
-
-        # Add optional parameters as query parameters if provided
-        query_params = {}
-        if "ExpectedBucketOwner" in kwargs:
-            query_params["expected-bucket-owner"] = kwargs["ExpectedBucketOwner"]
-
-        # Sign the request
-        signed_headers = self._presigner.sign_request_headers(
-            method="HEAD",
-            path=path,
-            headers=headers,
-        )
-
-        # Build the full URL
-        url = base_url + path
-        if query_params:
-            url = f"{url}?{self._build_query_string(query_params)}"
-
-        # Execute the request
+        url, signed_headers = self._prepare_head_bucket(Bucket, **kwargs)
         response = await self._execute_request("HEAD", url, signed_headers)
+        return self._parse_head_bucket_response(response, Bucket)
 
-        # Handle error responses
-        if response.status_code == 404:
-            raise NoSuchBucketError(
-                f"Bucket '{Bucket}' does not exist or is not accessible"
-            )
-        elif response.status_code == 403:
-            raise NoSuchBucketError(
-                f"Access denied to bucket '{Bucket}'. Check credentials and permissions."
-            )
-        elif response.status_code == 400:
-            raise PresignError(f"Bad request to bucket '{Bucket}': {response.text}")
-        elif response.status_code != 200:
-            raise PresignError(
-                f"HEAD request failed with status {response.status_code}: {response.text}"
-            )
-
-        # Extract metadata from response headers
-        bucket_region = response.headers.get("x-amz-bucket-region", self.region)
-
-        return {
-            "BucketRegion": bucket_region,
-            "ResponseMetadata": {
-                "HTTPStatusCode": response.status_code,
-                "HTTPHeaders": dict(response.headers),
-            },
-        }
-
-    async def head_object(self, Bucket: str, Key: str, **kwargs):
+    async def head_object(self, Bucket: str, Key: str, **kwargs) -> dict[str, Any]:
         """Check if an object exists and retrieve its metadata.
 
         Performs a HEAD request to retrieve object metadata without downloading the object.
@@ -549,74 +302,11 @@ class AsyncClient:
             1024
 
         """
-        if not Bucket:
-            raise PresignError("Missing required parameter 'Bucket'")
-        if not Key:
-            raise PresignError("Missing required parameter 'Key'")
-
-        # Build the request
-        base_url, path, headers = self._build_request_url(Bucket, Key)
-
-        # Add optional parameters as query parameters if provided
-        query_params = {}
-        if "VersionId" in kwargs:
-            query_params["versionId"] = kwargs["VersionId"]
-
-        # Sign the request
-        signed_headers = self._presigner.sign_request_headers(
-            method="HEAD",
-            path=path,
-            headers=headers,
-        )
-
-        # Build the full URL
-        url = base_url + path
-        if query_params:
-            url = f"{url}?{self._build_query_string(query_params)}"
-
-        # Execute the request
+        url, signed_headers = self._prepare_head_object(Bucket, Key, **kwargs)
         response = await self._execute_request("HEAD", url, signed_headers)
+        return self._parse_head_object_response(response, Bucket, Key)
 
-        # Handle error responses
-        if response.status_code == 404:
-            raise PresignError(
-                f"Object '{Key}' in bucket '{Bucket}' does not exist or is not accessible"
-            )
-        elif response.status_code == 403:
-            raise PresignError(
-                f"Access denied to object '{Key}' in bucket '{Bucket}'. Check credentials and permissions."
-            )
-        elif response.status_code == 400:
-            raise PresignError(f"Bad request for object '{Key}': {response.text}")
-        elif response.status_code != 200:
-            raise PresignError(
-                f"HEAD request failed with status {response.status_code}: {response.text}"
-            )
-
-        # Extract metadata from response headers
-        result = {
-            "ContentLength": int(response.headers.get("content-length", 0)),
-            "LastModified": response.headers.get("last-modified"),
-            "ETag": response.headers.get("etag"),
-            "ResponseMetadata": {
-                "HTTPStatusCode": response.status_code,
-                "HTTPHeaders": dict(response.headers),
-            },
-        }
-
-        # Add optional metadata if present
-        if "content-type" in response.headers:
-            result["ContentType"] = response.headers.get("content-type")
-        if "cache-control" in response.headers:
-            result["CacheControl"] = response.headers.get("cache-control")
-        if "content-encoding" in response.headers:
-            result["ContentEncoding"] = response.headers.get("content-encoding")
-        if "x-amz-version-id" in response.headers:
-            result["VersionId"] = response.headers.get("x-amz-version-id")
-
-        return result
-
-    async def create_bucket(self, Bucket: str, **kwargs):
+    async def create_bucket(self, Bucket: str, **kwargs) -> dict[str, Any]:
         """Create a new S3 bucket.
 
         Performs a PUT request to create a new S3 bucket.
@@ -646,66 +336,6 @@ class AsyncClient:
             '/mybucket'
 
         """
-        if not Bucket:
-            raise PresignError("Missing required parameter 'Bucket'")
-
-        # Build the request
-        base_url, path, headers = self._build_request_url(Bucket)
-
-        # Build request body for CreateBucketConfiguration if provided
-        body = b""
-        if "CreateBucketConfiguration" in kwargs:
-            config = kwargs["CreateBucketConfiguration"]
-            if "LocationConstraint" in config:
-                location = config["LocationConstraint"]
-                # Only send body if location is not us-east-1 (default)
-                if location and location != "us-east-1":
-                    body = (
-                        f"<CreateBucketConfiguration>"
-                        f"<LocationConstraint>{location}</LocationConstraint>"
-                        f"</CreateBucketConfiguration>"
-                    ).encode("utf-8")
-                    headers["Content-Type"] = "application/xml"
-
-        # Sign the request with the body
-        signed_headers = self._presigner.sign_request_headers(
-            method="PUT",
-            path=path,
-            headers=headers,
-            body=body,
-        )
-
-        # Build the full URL
-        url = base_url + path
-
-        # Execute the request
+        url, signed_headers, body = self._prepare_create_bucket(Bucket, **kwargs)
         response = await self._execute_request("PUT", url, signed_headers, body)
-
-        # Handle error responses
-        if response.status_code == 409:
-            # Check if it's BucketAlreadyExistsError or BucketAlreadyOwnedByYouError
-            if "BucketAlreadyOwnedByYouError" in response.text:
-                raise BucketAlreadyOwnedByYouError(
-                    f"Bucket '{Bucket}' already exists and is owned by you"
-                )
-            else:
-                raise BucketAlreadyExistsError(f"Bucket '{Bucket}' already exists")
-        elif response.status_code == 400:
-            raise PresignError(
-                f"Bad request to create bucket '{Bucket}': {response.text}"
-            )
-        elif response.status_code not in (200, 201):
-            raise PresignError(
-                f"PUT request failed with status {response.status_code}: {response.text}"
-            )
-
-        # Extract location from response headers or construct it
-        location = response.headers.get("Location", f"/{Bucket}")
-
-        return {
-            "Location": location,
-            "ResponseMetadata": {
-                "HTTPStatusCode": response.status_code,
-                "HTTPHeaders": dict(response.headers),
-            },
-        }
+        return self._parse_create_bucket_response(response, Bucket)
